txscript: Optimize alt stack drop.
Profiling shows that approximately 0.67% of the total allocations are
coming from DropN even though it should realistically be zero.

The cause is that DropN returns an error if it is called with 0 and the
code that drops the alternate data stack in between scripts calls it
with the depth of the alternate data stack, which is almost always zero.

The call has no observable effect on the result of the script execution
in that case since the error is intentionally ignored, there is nothing
to be dropped, and is therefore correct semantically, however, it is
wasteful to generate and throw away the error.

There really is no good reason for the internal stack implementation to
return an error from DropN when called with 0 since it equivalent to a
NOP.

I have done a careful analysis of all call sites to ensure the change
will break any observable semantics.

Thus, this modifies DropN to remove the unnecessary error condition and
updates the associated internal stack implementation test accordingly.