multi: Add simnet documentation and setup script.
This adds a script for simplifying simnet usage by creating a fresh
version of a preconfigured environment each time it is started along
with documentation that provides information regarding simnet, how to
make use of the script and environment, including some basic tasks and
common issues, as well as some information about manual configuration.

This script is something that has evolved over the years from one I
originally wrote and was later modified by Matheus Degiovani to make use
of tmux.  The version provided here also contains some new features and
convenience scripts that were written specifically to help newer users
and simplify common tasks such as performing a chain reorganization and
shutting down the environment.

Some of the documentation is based on a gist I created a long time ago,
which was then converted to the content currently on the documentation
site, but this is a much more polished version that includes the extra
information about the script and associated environment it provides and
is better served alongside the script in the repository.