blockchain: Remove superfluous blockidx fields.
Originally, some information that is related to vote ticket hashes and
revocations was added to the block index because it was needed prior to
the full node index being available and also because there was no
dynamic loading of the relevant stake node information in various
circumstances which meant the information had to be available on the
nodes at all times.

However, since that time, things have changed quite a bit such that the
full block index is now available in memory, the stake nodes and related
information are now efficiently loaded on demand, and stake node pruning
has been optimized accordingly.

As a result, it no longer makes sense to store all of this information.

Profiling shows this reduces the total inuse memory by about 19.5%.