blockchain: Migration for v3 block index.
This adds the code to migrate the database to a new v3 block index.

It involves removing the ticket hashes associated with vote info and
revocations that are no longer needed.  The process can be interrupted
at any point and future invocations will resume from the point it was
interrupted.