contrib: Update simnet script for dcrwallet master
The prompts when creating the wallet were changed when the wallet
passphrase was part of the application configuration, which caused the
script to hang.  This fixes the hang by replying to the additional
prompt, and updates the script to pipe input to dcrwallet's stdin to
reply to the prompts, rather than using tmux to send keys.