contrib: Support env var in simnet setup script.
This modifies the dcr_tmux_simnet_setup.sh script to set the root of all
of the nodes to the environment variable DCR_SIMNET_ROOT when it is set
and fallback to the existing path when it is not.

It also updates contrib/README.md and docs/simnet_environment.mediawiki
to document the new capability.