blockchain: Decentralized Treasury db migration.
This adds code to migrate the database to the new schema required by the
decentralized treasury implementation.

It involves updating the utxo set and spend journal entries to account
for the change of the position of the fully spent bit within the flags
from bit 4 to bit 6.  The process can be interrupted at any point and
future invocations will resume from the point it was interrupted.