main: Unexport main package exports.
This ensures that all code in the main package of root module is not
exported since none of it should be accessed outside of the main package
itself.

While here, it also updates the NAT discover func to avoid using named
variables which ensures error are not shadowed and eaten.