sampleconfig: Make constant a function instead.
This modifies the sampleconfig package to contain a single function
named FileContents instead of a constant.

This is being done because changing a constant is technically a major
semver breaking change while changing the contents returned by a
function is not and the goal is to ultimately make the root module
semver compliant.