peer: Handle InitState messages
This modifies the peer package to handle GetInitState/InitState
messages by adding the appropriate listeners. It also bumps the max
supported protocol version to version 8, which is needed to be able to
decode the messages.

Note this modifies the peer package's go.mod to use the local copy of
the wire module in order to have access to the new messages.