server: Send and respond to InitState msgs
This changes the server to support the most recent protocol version (v8
or InitStateVersion).

On this version, instead of using the getminingstate/miningstate pair of
messages, peers use getinitstate/initstate messages which allow for
finer grained control over what information is sent upon startup.

The older protocol versions are still supported so that the node may
continue to both request and serve initial mining state to older nodes
of the network.