blockchain: Add test func to remove deployment.
This adds a function to the common test infrastructure to remove
deployments for a given vote ID from parameters for use in testing cases
where it is useful to ensure behavior is correct on networks for which
an agenda is already active before heights that are not possible to hit
when the agenda has to be voted in.

Note that it is intentionally unused at this point, but future commits
will make use of it.