rpcserver: Add support for gettreasuryspendvotes
The new gettreasuryspendvotes rpc command is useful to keep track of the
voting progress of treasury spend transactions as well as to query for
the status of old tspends.

When invoked without any parameters, the command tallies the vote counts
for all tspends currently in the node's mempool up to the current tip
block.

The first parameter ("Blocks") can be used to limit the block up to
which the tallies are performed, which is useful for historical queries.

The second parameter ("TSpends") can be used to narrow down the list of
tspends that should be queried. It can be used to query for tspends that
have already been mined.

A set of unit tests ensures proper behavior for the new command.