chaincfg: Add checkpoints for upcoming release.
This adds checkpoints at the following heights:

mainnet:

- 237310
- 277940
- 318020
- 340120
- 362120
- 404170
- 424170
- 444170
- 464170
- 483600

testnet:

- 148320
- 213120
- 347140
- 411940
- 476740
- 515730

Note that this also includes some historical checkpoints to limit the
size of the gaps between each checkpoint to be around 3 months max.
This is being done because recent optimizations have sped up processing
to the point that larger gaps can trigger peers to get banned during IBD
in some cases.