rpc/jsonrpc/types: Prepare v2.1.0.
This updates the rpc/jsonrpc/types module dependencies, the copyright
year in the files modified since the previous release, and serves as a
base for rpc/jsonrpc/types/v2.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrjson@v3.1.0

The full list of updated direct dependencies since the previous
rpc/jsonrpc/types/v2.0.0 release are as follows:

- github.com/decred/dcrd/dcrjson@v3.1.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.