dcrutil: Prepare v3.0.0.
This updates the dcrutil module dependencies and serves as a base for
dcrutil/v3.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/ripemd160@v1.0.1
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.1
- github.com/decred/dcrd/dcrec/secp256k1/v3@v3.0.0

The full list of updated direct dependencies since the previous
dcrutil/v2.0.1 release are as follows:

- github.com/decred/base58@v1.0.3
- github.com/decred/dcrd/crypto/ripemd160@v1.0.1
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.1
- github.com/decred/dcrd/dcrec/secp256k1/v3@v3.0.0
- github.com/decred/dcrd/wire@v1.4.0

Further, the following dependencies that were previously required in the
dcrutil/v2.0.1 release are no longer required:

- github.com/decred/dcrd/chaincfg/v2

Finally, all modules in the repository that depend on dcrutil are tidied
to ensure they are updated to use the latest versions hoisted forward as
a result.