rpcserver: Forward HTTP server err msgs to logger.
The standard library HTTP server writes all errors using standard error
by default when and error logger is not configured.  Currently, an error
logger is not being set which means the errors completely bypass the
logging system.

Since the required error logger must be from the standard library
logging package and a custom logger is used instead, this creates a shim
to forward the log messages from a standard lib logger to the RPC server
logger and configures the HTTP server to use it accordingly.