database: Prepare v2.0.2.
This updates the database module dependencies and serves as a base for
database/v2.0.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/v3@v3.0.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/slog@v1.1.0
- github.com/syndtr/goleveldb@v1.0.1-0.20200815110645-5c35d600f0ca

The full list of updated direct dependencies since the previous
dcrutil/v2.0.1 release are as follows:

- github.com/decred/dcrd/chaincfg/v3@v3.0.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/dcrd/wire@v1.4.0
- github.com/decred/slog@v1.1.0
- github.com/syndtr/goleveldb@v1.0.1-0.20200815110645-5c35d600f0ca

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.