txscript: Prepare v3.0.0.
This updates the txscript module dependencies and serves as a base for
txscript/v3.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/v3@v3.0.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/slog@v1.1.0

The full list of updated direct dependencies since the previous
txscript/v2.1.0 release are as follows:

- github.com/dchest/siphash@v1.2.1
- github.com/decred/dcrd/chaincfg/v3@v3.0.0
- github.com/decred/dcrd/crypto/ripemd160@v1.0.1
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.1
- github.com/decred/dcrd/dcrec/secp256k1/v3@v3.0.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/dcrd/wire@v1.4.0
- github.com/decred/slog@v1.1.0

Finally, all modules in the repository that depend on txscript are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.