blockchain/standalone: Remove txscript dep.
The purpose of the standalone module is to avoid extra dependencies.
The treasury work unfortunately went against that goal and introduced a
dependency on txscript.  This remedies that by defining the necessary
opcodes directly which is perfectly acceptable because they are used in
consensus code and thus can't be changed without a vote.