mining: Remove unneeded disapproval check
This removes a now unneeded check for unmineable transactions due to
block disapproval from the template generation code.

Previously this check was meant to ensure that the list of transactions
considered for inclusion in a block template did not include
transactions with an input referencing a regular transaction from a
parent disapproved block.

However this check is redudant due to check executed earlier in the
template building loop that asserts all inputs are valid utxos
according to the PoV of the parent block and taking into account whether
the new block being created will approve or disapprove it.

The check being removed had a bug due to the dcrutil.Tx being duplicated
without also copying the internal tree, which caused transactions to
later fail to be added to the template due to an unknown tree being
used.