blockchain: Simplify old block ver upgrade checks.
This converts the series of checks that rejects old block versions once
a majority of the network has upgraded to a loop.

When there were only a few this didn't matter, but as the number grows,
it is nicer to only need to change the number in a single place.