rpcserver: Correct known wallet method handling.
This corrects the code to return the expected RPC error in the case the
requested method is one known to be a wallet method or known to be
intentionally unimplemented.

Since dcrjson was updated to work with types instead of strings, parsing
a method that is not registered now fails, so the special error returns
are now handled there as opposed to the old approach of adding handlers
for certain method names which just returned the appropriate error.

This allows the standardCmdResult method to be simplified since it no
longer needs to include the logic for dealing with special handlers and
therefore it is updated accordingly.

Finally, the special handlers are removed since they are no longer
needed.