gcs: Prepare v2.1.0.
This updates the gcs module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
gcs/v2.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/blockchain/stake/v3@v3.0.0

The full list of updated direct dependencies since the previous
gcs/v2.0.1 release are as follows:

- github.com/decred/dcrd/blockchain/stake/v3@v3.0.0
- github.com/decred/dcrd/txscript/v3@v3.0.0
- github.com/decred/dcrd/wire@v1.4.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.