peer: Prepare v2.2.0.
This updates the peer module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
peer/v2.2.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/lru@v1.1.0
- github.com/decred/dcrd/txscript/v3@v3.0.0

The full list of updated direct dependencies since the previous
peer/v2.1.0 release are as follows:

- github.com/decred/dcrd/lru@v1.1.0
- github.com/decred/dcrd/txscript/v3@v3.0.0
- github.com/decred/dcrd/wire@v1.4.0
- github.com/decred/slog@v1.1.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.