rpcclient: Prepare v6.0.0.
This updates the rpcclient module dependencies, the copyright year in
the files modified since the previous release, and serves as a base for
rpcclient/v6.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrjson/v3@v3.1.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/dcrd/gcs/v2@v2.1.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v2@v2.1.0

The full list of updated direct dependencies since the previous
rpcclient/v5.0.1 release are as follows:

- github.com/decred/dcrd/dcrjson/v3@v3.1.0
- github.com/decred/dcrd/dcrutil/v3@v3.0.0
- github.com/decred/dcrd/gcs/v2@v2.1.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v2@v2.1.0
- github.com/decred/dcrd/wire@v1.4.0
- github.com/decred/slog@v1.1.0
- github.com/gorilla/websocket@v1.4.2

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.