[release-v1.6] mining: Prevent panic in child prio item handling.
This prevents an issue with the way that children are added to the
priority queue when building mining templates.  In particular, it
ensures the child has already had priority item built for it before
adding to the prioirty queue since it is possible that the child has not
yet been seen under certain circumstances such as when there are
complicated transaction graphs of unconfirmed transactions.