blockchain: Fix stxosToScriptSource
This fixes the stxosToScriptSource function so that it takes into
account the new treasury transactions which don't have any inputs
(treasury bases and treasury spends).

Previously, the presence of one of this transactions in the blockchain
would cause a panic during addrindex calculation due to previous scripts
for inputs of those transactions not existing.