blockchain: Make ver 5 to 6 db upgrades work again.
This allows version 5 to 6 database upgrades to work again by
implementing code that is independent of the main code and works with
the data structures as they existed in the version 5 database.

While here, it also updates some other cases in both older and newer
upgrade code to make the upgrades stable even if the primary code
changes.

An high-level overview of the key changes is:

- Clear the failed validation status of the v2 block index entries
  directly instead of loading the code into the main index, modifying
  it, and flushing it
- Avoid using the unversioned main funcs related to serializing and
  deserializing the spend journal
- Implement code to directly create a script source from a v2 spend
  journal entry
- Reimplement the migration code that initializes the v2 GCS filters to
  independently read the block tree and discover the main chain