blockchain: Clear failed block flags for HF.
Since the next release of the software will include a vote to change the
consensus rules, this unmarks all blocks previously marked as having
failed validation so they are eligible for validation again under what
will likely become new consensus rules.  This ensures clients that did
not update prior to new rules activating are able to automatically
recover under the new rules without having to download the entire chain
again.

This really should have been done as a part of the version 7 upgrade,
but it wasn't, so this bumps the database version accordingly.

Also, since the version has to be bumped, this also takes the
opportunity to roll the fix to ensure the treasury buckets that weren't
created in the v7 upgrade in the original RC into the v8 upgrade.