rpcclient: Prepare v6.0.1.
This updates the rpcclient module dependencies and serves as a base for
rpcclient/v6.0.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/rpc/jsonrpc/types/v2@v2.2.0

The full list of updated direct dependencies since the previous
rpcclient/v6.0.0 release are as follows:

- github.com/decred/dcrd/rpc/jsonrpc/types/v2@v2.2.0

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.