blockchain: Simplify upgrade single run stage code.
This refactors the upgrade code that avoids running a single stage again
in the case of interrupted upgrades to consolidate the logic into a
single function that accepts a callback and key name to use for tracking.

The net effect is that the individual functions that handle whatever
migration is necessary by that stage can focus specifically on that
logic and the caller then handles whether or not to apply the stage
completion tracking by simply wrapping it with the new func.

This also allows a case of tracking that wasn't actually needed but was
originally made a part of the migration function since it is intended to
be called by future upgrades which might consist of multiple stages.

The updated approach handles that type of situation much more cleanly.