blockchain: Simplify upgrade batching logic.
This introduces a new function for handling repeatedly calling a
provided batch function which, as the name implies, performs database
updates and migrations in batches.

The logic for handling this is currently used several times throughout
the upgrade code and will very likely continue to be used in the future,
so, rather than repeating it multiple times with slight variations as is
currently required, the approach introduced here allows the batch
function to signal when it should no longer be invoked due to finishing
the final batch and passed into the new function which handles the
common transaction updating and looping logic.

Initially, only the migration function for clearing the v2 block index
flags is updated to use the new approach.  The remaining funcs will be
updated in future commits.