[release-v1.6] gencerts: Use the P-256 curve by default.
This modifies the default curve used to generate the certificates by the
gencerts utility to be P-256 instead of P-521 to match a recent similar
change made to the RPC server automatic generation logic.

Unfortunately, Chromium removed support for P-521 and since that is what
electron uses under the hood, Decrediton can no longer connect to dcrd
RPC servers using a certificate with P-521.

Thus, given the gencerts utility is used in various scenarios which
involve interop with the aforementioned software, update it as well.