rpcserver: Correct getpeerinfo for peers w/o conn.
This corrects the getpeerinfo RPC handler to avoid a panic when the
local address is nil which is the case when a peer is not currently
connected.  This can happen for cases such as permanent peers and during
process shutdown as peers are disconnected and being cleaned up.

While here, it also updates the comments to specifically call out that
the LocalAddr method returns nil when a peer is not connected to better
inform callers they are expected to handle that condition.