[release-v1.6] rpcserver: Correct treasury vote status handling.
This corrects several of the RPC handlers to base the status of whether
or not the decentralized treasury rules are active based on the location
of the element in question versus whatever the current tip happens to
be.

In particular, the following RPCs are corrected:

- getblock
- getblocksubsidy
- getrawtransaction
- gettxout
- rescan
- searchrawtransactions
- ticketfeeinfo
- txfeeinfo

While here, it also modifies searchrawtransaction to use the height in
the header instead of performing another lookup from chain since it is
already available there and removes a TODO that no longer applies.