[release-v1.6] multi: Add tx inputs treasurybase RPC support.
This adds a new field to the Vin and VinPrevOut fields of json/rpc/types
to identify if the input is a treasurybase or not and updates both the
marshalling code as well as the relevant rpcserver results accordingly.
Since a treasurybase does not have a signature script like a coinbase
and stakebase, the new field is a boolean versus a string like the
others.

In addition, it updates the rpcserver help to document the new field.

At a high level, this means the verbose output of decoderawtransaction,
getblock, getrawtransaction, searchrawtransactions, and websocket
transaction accepted notifications will now identify treasurybase
inputs.

Finally, it bumps the rpcserver to version 6.2.0 to signal support for
the new features.