[release-v1.6] multi: Add tx inputs treasuryspend RPC support.
This adds a new field to the Vin and VinPrevOut fields of json/rpc/types
to identify if the input is a treasury spend or not and updates both the
marshalling code as well as the relevant rpcserver results accordingly.

In addition, it updates the rpcserver help to document the new field.

At a high level, this means the verbose output of decoderawtransaction,
getblock, getrawtransaction, searchrawtransactions, and websocket
transaction accepted notifications will now identify treasury spend
inputs.

The rpcserver version is not bumped since there have been no new
releases and release candidates since the previous one.